+++ cloud-siem-20220616/alibabacloud_cloud_siem20220616/models/_get_data_storage_request.py
@@ -1,0 +1,56 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class GetDataStorageRequest(DaraModel):
+    def __init__(
+        self,
+        lang: str = None,
+        region_id: str = None,
+        role_for: int = None,
+    ):
+        # The language of the response. Valid values:
+        # - **zh** (default): Chinese.
+        # - **en**: English.
+        self.lang = lang
+        # The region where the threat analysis data management center is located. Specify the management center based on the region of your assets. Valid values:
+        # - cn-hangzhou: the asset is in the Chinese mainland.
+        # - ap-southeast-1: the asset is outside China.
+        self.region_id = region_id
+        # The user ID of the member to which the administrator switches the view.
+        self.role_for = role_for
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.lang is not None:
+            result['Lang'] = self.lang
+
+        if self.region_id is not None:
+            result['RegionId'] = self.region_id
+
+        if self.role_for is not None:
+            result['RoleFor'] = self.role_for
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Lang') is not None:
+            self.lang = m.get('Lang')
+
+        if m.get('RegionId') is not None:
+            self.region_id = m.get('RegionId')
+
+        if m.get('RoleFor') is not None:
+            self.role_for = m.get('RoleFor')
+
+        return self
+